+++ CHANGELOG.md
@@ -10,6 +10,10 @@ x.y.z Release notes (yyyy-MM-dd)
 
 * Combine IDKit setup with PACECloudSDK setup. `IDKit.setup(...)` is no longer accessible. The `IDKit` is now initialized via `PACECloudSDK.setup(...)`. Therefore an optional `OIDConfiguration` parameter was added to the `Configuration` class of the `PACECloudSDK`, which must be initialized with at least the `clientId` and the `redirectUri`.
 
+### Fixes
+
+* Adjust handling of utm parameters
+
 ### Internal
 
 * Adjust TOTP secret handling